requirements.txt
fastapi>=0.110,<1.0
uvicorn>=0.29,<1.0
sqlalchemy>=2.0,<3.0
python-jose[cryptography]>=3.3
passlib[bcrypt]>=1.7
python-multipart>=0.0.6
pydantic-settings>=2.0
pytest>=7.0,<9.0
httpx>=0.25,<1.0
